third_party/blink/web_tests/external/wpt/url/url-setters.any_include=javascript-expected.txt

This is a testharness.js-based test.
[FAIL] URL: Setting <javascript://x/>.username = 'wario'
  assert_equals: expected "javascript://wario@x/" but got "javascript://x/"
[FAIL] URL: Setting <javascript://x/>.password = 'bowser'
  assert_equals: expected "javascript://:bowser@x/" but got "javascript://x/"
[FAIL] URL: Setting <javascript://x/>.port = '12'
  assert_equals: expected "javascript://x:12/" but got "javascript://x/"
Harness: the test ran to completion.
